debian: Avoid linking with -Bsymbolic-functions
Our tests want to be able to change symbols from the library and
even by using --wrap=XYZ it is not possible right now. One option
would be to use static linking but that is not always enabled, the
other is to skip tests on Ubuntu and the third one is to disable
that linking mode. This means that the dynamic linker needs to
spend some more time but we do spend this time on all other distros
and this looks acceptable.