fsm: explicitly log removal from parent fsm
Since removing an FSM from its parent twice causes a segfault, it is very
interesting to see when that is attempted.

Removing could be made more robust, but logging is interesting for
investigating why an FSM is being removed twice in the first place (currently
the case in openbsc's vlr_lu_fsm).

Change-Id: Idec6b7aa5344f1e903c9d2aa2a3640cab0d70fb0